# central-server/openapi-generator/openapi.yaml
openapi: 3.1.0
info:
  title: Smart Attendance System API
  version: "1.0.0"
  description: This is the API documentation for the Smart Attendance System.
  # license:
  #   name: Apache 2.0
  #   url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://smartattendance.fastwebcm.local/api/v1
    description: Core Business Logic Endpoint
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid input
    Unauthorized:
      description: Unauthorized - Invalid or missing token
    NotFound:
      description: Resource not found
    InternalServerError:
      description: Internal server error
  schemas:
    UserType:
      type: string
      enum: [student, staff]
    UserStatus:
      type: string
      enum: [active, inactive, dismissed]
    BiometricEnrollmentStatus:
      type: string
      enum: [pending, completed]
...
    PaginationMeta:
      type: object
      required:
        - total_records
        - current_page
        - total_pages
        - limit
      properties:
        total_records:
          type: integer
          description: Total quantity of active users matching structural selection metrics.
          example: 142
        current_page:
          type: integer
          description: The currently pulled response page offset iteration.
          example: 1
        total_pages:
          type: integer
          description: Maximum aggregate count of dynamic pagination user slices available.
          example: 3
        limit:
          type: integer
          description: Total boundary count constraint cap assigned for user blocks.
          example: 50
...
    User:
      type: object
      required:
        - fname
        - lname
        - user_type
      properties:
        id:
          type: integer
        fname:
          type: string
          example: Ichami
        lname:
          type: string
          example: Brandon
        email:
          type: ["string", "null"]
          example: brandon@example.com
        gender:
          type: string
          enum: [male, female]
        user_type:
          $ref: "#/components/schemas/UserType"
        status:
          type: ["string", "null"]
        biometric_enrollment_status:
          $ref: "#/components/schemas/BiometricEnrollmentStatus"
        role_id:
          type: integer
          description: role id for privileged staff
        class_id:
          type: integer
          description: current class of the student
        regno:
          type: string
        username:
          type: string
        password:
          type: string
    UserCreate:
      allOf:
        - $ref: "#/components/schemas/User"
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        fname:
          type: string
        lname:
          type: string
        name:
          type: string
        email:
          type: string
          example: brandon@example.com
        gender:
          type: string
          enum: [male, female]
        status:
          $ref: "#/components/schemas/UserStatus"
        biometric_enrollment_status:
          $ref: "#/components/schemas/BiometricEnrollmentStatus"
        role:
          type: string
        class:
          type: string
        user_type:
          $ref: "#/components/schemas/UserType"
        regno:
          type: string
        username:
          type: string
        class_id:
          type: integer
        role_id:
          type: integer
    PaginatedUsersResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    StudentInput:
      type: object
      required: [id, sregnum, fname, lname, gender, cname]
      properties:
        id:
          type: integer
          description: The online database studentid.
          example: 1445
        sregnum:
          type: string
          description: Unique registration number.
          example: "STU-2026-A1B2"
        fname:
          type: string
          example: "John"
        lname:
          type: string
          example: "Doe"
        gender:
          type: string
          enum: [Male, Female, Other]
          example: "Male"
        cname:
          type: string
          description: The class name string.
          example: "Form 5 Science"
    StaffInput:
      type: object
      required: [id, tregnum, fname, lname]
      properties:
        id:
          type: integer
          description: The online database teacherid.
          example: 82
        tregnum:
          type: string
          description: Unique teacher registration number.
          example: "STF-2026-X8B2"
        fname:
          type: string
          example: "Jane"
        lname:
          type: string
          example: "Smith"
    SyncRequest:
      type: object
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/StudentInput"
        staff:
          type: array
          items:
            $ref: "#/components/schemas/StaffInput"
    SyncResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        studentIds:
          type: array
          description: List of student IDs successfully processed to be sent back to online ACK.
          items:
            type: integer
          example: [1445, 1446]
        staffIds:
          type: array
          description: List of teacher IDs successfully processed to be sent back to online ACK.
          items:
            type: integer
          example: [82]
    # Role Schema
...
paths:
...
  /api/v1/users:
    post:
      operationId: createUser
      summary: Create a new user
      tags: [Users]
      requestBody:
        required: true
        description:
          Ensure you insert into staff tbl specifying the role and of the
          staff if the user type is staff
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      operationId: listUsers
      summary: List all users
      tags: [Users]
      parameters:
        - in: query
          name: user_type
          schema:
            type: ["string", "null"]
          description: Filter by user type
        - in: query
          name: status
          schema:
            type: ["string", "null"]
          description: Filter by status
        - in: query
          name: search
          schema:
            type: string
          description: Contextual string match across names or student registration numbers
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: The pagination page index number to view
        - in: query
          name: role
          schema:
            type: ["string", "null"]
        - in: query
          name: class_id
          schema:
            type: integer
          description: Filter users by numeric class ID
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Maximum items returned per pagination block page window
      responses:
        "200":
          description: Paginated matrix list of system users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsersResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/users/sync:
    post:
      operationId: syncUsers
      summary: Sync users from the online SSEC database
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRequest"
      responses:
        "200":
          description: Users synced successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/users/{id}:
    get:
      operationId: getUserById
      summary: Get a user by ID
      tags: [Users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The unique ID of the user
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      operationId: updateUser
      summary: Update a user
      tags: [Users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The unique ID of the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      operationId: deleteUser
      summary: Delete a user
      tags: [Users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...